switch form from log-in to sign-up
