+++ client/src/pages/SignUp.tsx
@@ -1,8 +1,9 @@
+import LoginSignupStructure from "../components/LoginSignupStructure";
+
+
 const SignUp = () => {
     return (
-        <h1>Sign Up
-        
-        </h1>
+        <LoginSignupStructure/>
     )
 };
